Added content to the Spells page (HTML, CSS, JS)
I created the basic html for the Spells page and added some styles. I called and API to pull and display the D&D cantrips and spells, only cantrips are visible for now.

Also added a pop-up view of the spell card.

Note: Start to work on the search bar functionality of this page.